IB/core: Add kref to IB devices
Previously. we used device_mutex lock in order to protect
the device's list. That means that in order to guarantee a
device isn't freed while we use it, we had to lock all
devices.

Adding a kref per IB device. Before an IB device
is unregistered, we wait before its not held anymore.

Change-Id: Iabab2780c3905079d3869f0f43e42ef49b70d7d3
Signed-off-by: Matan Barak <matanb@mellanox.com>