IB/core: Add ib_find_cached_gid_by_port cache API
Sometime consumers might want to search for a GID in a specific port.
For example, when a WC arrives and we want to search the GID
that matches that IP - it's better to search only the relevant
port.
Adding an API to find a GID by an optionally netdevice only
on the relevant port.

Change-Id: If3f4eb83bc850aac8c17e23021a8e33e55f79d44
Signed-off-by: Matan Barak <matanb@mellanox.com>